requirements.txt
cachetools==5.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.11.0
django-heroku==0.3.1
django-import-export==2.5.0
et-xmlfile==1.1.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
gunicorn==20.0.4
httplib2==0.22.0
idna==2.10
inflect==5.0.2
MarkupPy==1.14
mysqlclient==2.0.3
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.9
pandas==2.0.3
psycopg2==2.8.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2020.5
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tablib==3.1.0
tzdata==2023.3
urllib3==1.26.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
